Move misplaced constants out of layer-violating directories
Pure refactor — no behavior changes. Three constants relocated to fix
layer inversions identified in a recent audit:

- SESSION_STATE_FILENAME: src/docker/pty-types.ts → src/config/paths.ts.
  Removes the config/paths.ts ↔ docker/pty-types.ts cycle (config was
  importing from docker for a session filename).

- JOB_ID_PATTERN: src/cron/types.ts → src/types/slug.ts. Config no
  longer reaches into cron; the constant is just SLUG_PATTERN aliased
  for clarity, so it sits next to its source.

- RESERVED_SERVER_NAMES: src/docker/proxy-tools.ts →
  src/types/reserved-servers.ts (new file). Both src/config/index.ts
  and src/trusted-process/policy-roots.ts had to import from docker/
  for a system-wide concept; they now share a neutral home.

All exported names unchanged. Importers updated. Lint, typecheck, and
full test suite (4283 passed) all green.
